+++ resources/scripts/components/admin/servers/overview/ServerOverviewContainer.tsx
@@ -1,7 +1,20 @@
 import ServerContentBlock from '@/components/admin/servers/ServerContentBlock'
+import Button from '@/components/elements/Button'
+import { AdminServerContext } from '@/state/admin/server'
+import { Link } from 'react-router-dom'
 
 const ServerOverviewContainer = () => {
-    return <ServerContentBlock title={'Overview'}></ServerContentBlock>
+    const server = AdminServerContext.useStoreState(state => state.server.data!)
+
+    return (
+        <ServerContentBlock title={'Overview'}>
+            <Link to={`/servers/${server.id}`}>
+                <Button className='inline-flex items-center' as={'a'}>
+                    Enter Server Console
+                </Button>
+            </Link>
+        </ServerContentBlock>
+    )
 }
 
 export default ServerOverviewContainer