+++ resources/scripts/components/servers/overview/ServerOverviewContainer.tsx
@@ -1,15 +1,19 @@
+import ServerAdminBlock from '@/components/servers/overview/ServerAdminBlock'
 import ServerDetailsBlock from '@/components/servers/overview/ServerDetailsBlock'
 import ServerPowerBlock from '@/components/servers/overview/ServerPowerBlock'
 import ServerTerminalBlock from '@/components/servers/overview/ServerTerminalBlock'
 import ServerContentBlock from '@/components/servers/ServerContentBlock'
+import { useStoreState } from '@/state'
 
 const ServerOverviewContainer = () => {
+  const rootAdmin = useStoreState(state => state.user.data!.rootAdmin)
   return (
     <ServerContentBlock title='Overview'>
       <ServerPowerBlock />
       <div className='grid grid-cols-10 gap-6'>
         <ServerDetailsBlock />
         <ServerTerminalBlock />
+        {rootAdmin && <ServerAdminBlock />}
       </div>
     </ServerContentBlock>
   )